feat(ui): moved wine/dxvk settings to separate page
